+++ src/api_entreprise/models/certification_qualibat.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+from ..utils.metaclasses import _AddMarshmallowSchema
+
+
+@dataclass
+class CertificationQualibat(metaclass=_AddMarshmallowSchema):
+    document_url: str
+    expires_in: int
+    """Nombre de secondes avant expiration"""