+++ tests/test_certification_rge.py
@@ -8,13 +8,13 @@ from .fixtures import vcr_folder
 
 
 @vcr.use_cassette(f"{vcr_folder}/certif_rge.megalis.yml")
-def test_raw_ca(api: ApiEntreprise):
+def test_raw_certif_rge(api: ApiEntreprise):
     certif = api.raw_certification_rge(25351449100047)
     assert certif is not None
 
 
 @vcr.use_cassette(f"{vcr_folder}/certif_rge.megalis.yml")
-def test_ca(api: ApiEntreprise):
+def test_certif_rge(api: ApiEntreprise):
     certif = api.certifications_rge(25351449100047)
     assert certif is not None
     assert isinstance(certif, list)