+++ CHANGELOG.md
@@ -3,16 +3,18 @@
 Changelog basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 et ce projet adhère au [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.1.0]
 
 ### Added
 
 - Ajout certification RGE
+- Ajout certification Qualibat
 
 ### Changed
 
 - Renommage model
 - handler qui groupe plusieurs autres handlers
+- refactoring [api.py](./src/api_entreprise/api.py)
 
 ### Fixed